requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2-binary==2.9.7
django-pgvector==0.2.4
celery==5.3.4
redis==5.0.1
sentence-transformers==2.2.2
transformers==4.35.2
diffusers==0.24.0
torch==2.1.1
Pillow==10.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.1 